all: move AppLock struct to types/app
