+++ autoscale/autoscale.go
@@ -25,6 +25,7 @@ import (
 	"github.com/tsuru/tsuru/provision"
 	"github.com/tsuru/tsuru/provision/node"
 	"github.com/tsuru/tsuru/safe"
+	appTypes "github.com/tsuru/tsuru/types/app"
 )
 
 const (
@@ -267,7 +268,7 @@ func (a *Config) runScalerInNodes(prov provision.NodeProvisioner, pool string, n
 	evt.Logf("running scaler %T for %q: %q", scaler, provision.PoolMetadataName, pool)
 	customData.Result, err = scaler.scale(pool, nodes)
 	if err != nil {
-		if _, ok := err.(app.ErrAppNotLocked); ok {
+		if _, ok := err.(appTypes.ErrAppNotLocked); ok {
 			evt.Logf("aborting scaler for now, gonna retry later: %s", err)
 			return
 		}
